Fix of Issue #82
view/vifcon_gase was changed so it uses the same syntax as the other displayed values.

device/vifcon_gase was changed to only show nan values if the values for PP21, PP22 and DM21 are below 1e-5 AND NOT 0, to have the same convention as the other values  from vifcon gase.

The fix itself is applied in the base class.

Tested on NemoOne with real hardware.